requirements.txt
pytest==6.0.2
paho-mqtt==1.5.0
